Adding a base project and sharing some core files.
