+++ Project/Src/AddIns/ReSharper710/Core/StyleCopCoreFactory.cs
@@ -20,6 +20,7 @@ namespace StyleCop.ReSharper710.Core
     #region Using Directives
 
     using StyleCop.Diagnostics;
+    using SourceCodeFactory = ReSharperBase.Core.SourceCodeFactory;
 
     #endregion
