+++ Project/Src/AddIns/ReSharper710/ReSharper710.csproj
@@ -273,8 +273,6 @@
     <Compile Include="Core\DeclarationHeader.cs" />
     <Compile Include="Core\FileHeader.cs" />
     <Compile Include="Core\ProjectSettingsFactory.cs" />
-    <Compile Include="Core\SourceCodeFactory.cs" />
-    <Compile Include="Core\StringBasedSourceCode.cs" />
     <Compile Include="Core\StyleCopCoreFactory.cs" />
     <Compile Include="Core\StyleCopReferenceHelper.cs" />
     <Compile Include="Core\StyleCopRules.cs" />
@@ -422,6 +420,10 @@
     <Compile Include="Violations\StyleCopHighlightingWarning.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\..\..\ReSharperBase\ReSharperBase.csproj">
+      <Project>{16c1d117-776a-4ca8-a02a-7f5e1da4359a}</Project>
+      <Name>ReSharperBase</Name>
+    </ProjectReference>
     <ProjectReference Include="..\..\StyleCop\StyleCop.csproj">
       <Project>{AF7A7FEA-C07B-456A-B072-C7E795A6CBD4}</Project>
       <Name>StyleCop</Name>
