+++ Project/StyleCop.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 2013
-VisualStudioVersion = 12.0.21005.1
+VisualStudioVersion = 12.0.40629.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8FBEC870-B153-4917-A53A-9FB8B369FCA4}"
 	ProjectSection(SolutionItems) = preProject
@@ -55,6 +55,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReSharper710", "Src\AddIns\ReSharper710\ReSharper710.csproj", "{CA8CAC03-6349-44C0-9AC3-FA49AAB58B53}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReSharper800", "Src\AddIns\ReSharper800\ReSharper800.csproj", "{A101FCFE-0024-4B4A-84AB-AB9C0442E9AF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReSharperBase", "ReSharperBase\ReSharperBase.csproj", "{16C1D117-776A-4CA8-A02A-7F5E1DA4359A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -146,6 +148,10 @@ Global
 		{A101FCFE-0024-4B4A-84AB-AB9C0442E9AF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A101FCFE-0024-4B4A-84AB-AB9C0442E9AF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A101FCFE-0024-4B4A-84AB-AB9C0442E9AF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{16C1D117-776A-4CA8-A02A-7F5E1DA4359A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16C1D117-776A-4CA8-A02A-7F5E1DA4359A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16C1D117-776A-4CA8-A02A-7F5E1DA4359A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16C1D117-776A-4CA8-A02A-7F5E1DA4359A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
